requirements.txt
numpy==1.20.1
Pillow==9.3.0
